Rename Packages to actual panpan configs
