+++ pan-pan-backend/src/main/java/com/panpan/ReactAndSpringDataRestApplication.java
@@ -1,4 +1,4 @@
-package com.greglturnquist.payroll;
+package com.panpan;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
