+++ pan-pan-backend/src/main/java/com/panpan/controller/HomeController.java
@@ -1,4 +1,4 @@
-package com.greglturnquist.payroll;
+package com.panpan.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
