+++ pan-pan-backend/src/main/java/com/panpan/repository/UserRepository.java
@@ -1,6 +1,8 @@
-package com.greglturnquist.payroll;
+package com.panpan.repository;
 
 import org.springframework.data.repository.CrudRepository;
+
+import com.panpan.model.User;
 
 public interface UserRepository extends CrudRepository<User, Long> {
